+++ tests/testthat/test-hash64.R
@@ -158,4 +158,9 @@ test_that("runif64 replace=FALSE edge cases", {
   expect_length(r2, 5L)
   expect_length(unique(r2), 5L)
   expect_true(all(r2 >= 1 & r2 <= 1000))
+
+  r3 = runif64(20L, 1, 20, replace=FALSE)
+  expect_length(r3, 20L)
+  expect_length(unique(r3), 20L)
+  expect_true(all(r3 >= 1 & r3 <= 20))
 })